implement note group crud apis (#1)
* Create note group crud endpoints

* add note group note add remove api

* self review

+++ adapter/src/main/java/com/wispnote/backend/adapter/source/jpa/projection/SourceSummaryProjection.java
@@ -13,6 +13,6 @@ public interface SourceSummaryProjection {
     Instant getLastActivityAt();
 
     default SourceSummary toModel() {
-        return new SourceSummary(getSource().toModel(), getNoteCount().intValue(), getLastActivityAt());
+        return new SourceSummary(getSource().toModel(), getNoteCount(), getLastActivityAt());
     }
 }
