+++ application/src/main/java/com/wispnote/backend/application/source/model/SourceSummary.java
@@ -3,6 +3,6 @@ package com.wispnote.backend.application.source.model;
 import java.time.Instant;
 
 public record SourceSummary(Source source,
-                            Integer noteCount,
+                            Long noteCount,
                             Instant lastActivityAt) {
 }
